Fix linux_rusage_t layout: timeval fields int32_t → int64_t (AArch64)
On Linux AArch64, struct timeval uses `long` (64-bit), so ru_utime_sec/usec
and ru_stime_sec/usec must be int64_t. Old layout was 128 bytes (wrong);
corrected layout is 144 bytes, matching the kernel struct rusage exactly.

wait4 (LINUX_NR_wait4=260) binding was already present; this fixes the
rusage struct written to user-space. Also update CLAUDE.md: add explicit
"manca da implementare" checklist for M11-05a with stub/passthrough/medium
breakdown and mark wait4 fix as done.

+++ include/linux_compat.h
@@ -160,10 +160,10 @@ typedef struct {
 } linux_utsname_t;
 
 typedef struct {
-    int32_t  ru_utime_sec;
-    int32_t  ru_utime_usec;
-    int32_t  ru_stime_sec;
-    int32_t  ru_stime_usec;
+    int64_t  ru_utime_sec;   /* Linux AArch64: long = 64-bit */
+    int64_t  ru_utime_usec;
+    int64_t  ru_stime_sec;
+    int64_t  ru_stime_usec;
     int64_t  ru_maxrss;
     int64_t  ru_ixrss;
     int64_t  ru_idrss;
